+++ EAServerStatus/Models/Notification.cs
@@ -6,13 +6,13 @@ namespace EAServerStatus.Models
 {
     public class Notification
     {
-        private readonly NotifyIcon notifyIcon;
+        private readonly NotifyIcon _notifyIcon;
 
         public Notification()
         {
-            notifyIcon = new NotifyIcon();
-            notifyIcon.Icon = Icon.ExtractAssociatedIcon(Assembly.GetExecutingAssembly().Location);
-            notifyIcon.BalloonTipClosed += (s, e) => notifyIcon.Visible = false;
+            _notifyIcon = new NotifyIcon();
+            _notifyIcon.Icon = Icon.ExtractAssociatedIcon(Assembly.GetExecutingAssembly().Location);
+            _notifyIcon.BalloonTipClosed += (s, e) => _notifyIcon.Visible = false;
         }
 
         private void Show(string title, string message, ToolTipIcon icon)
@@ -20,8 +20,8 @@ namespace EAServerStatus.Models
             if (string.IsNullOrEmpty(title)) title = " ";
             if (string.IsNullOrEmpty(message)) message = " ";
 
-            notifyIcon.Visible = true;
-            notifyIcon.ShowBalloonTip(3000, title, message, icon);
+            _notifyIcon.Visible = true;
+            _notifyIcon.ShowBalloonTip(3000, title, message, icon);
         }
 
         public void ShowInfo(string title, string message = "")