+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### v1.1.1 (3-Sep-2022)
+- Changed Offline status to Maintenance
+- Added data correction in chart
+- Refactored
 ### v1.1.0 (1-Sep-2022)
 - Chart bug fix & improvement
 - Added Data Error & Server Error status